cloud_functions/process/requirements.txt
functions-framework==3.4.0
flask==2.3.2
google-cloud-error-reporting==1.9.2
google-cloud-logging==3.6.0
google-cloud-resource-manager==1.10.2
jsonpath-ng==1.5.3
google-generativeai==0.4.1
vertexai==1.46.0
google-cloud-secret-manager==2.19.0
